move deposit data file to testutils
