+++ types/spectest/tests/beacon/deposit_data.go
@@ -3,6 +3,7 @@ package beacon
 import (
 	"encoding/hex"
 	"github.com/ssvlabs/ssv-spec/types"
+	"github.com/ssvlabs/ssv-spec/types/testingutils"
 	comparable2 "github.com/ssvlabs/ssv-spec/types/testingutils/comparable"
 	"github.com/stretchr/testify/require"
 	"reflect"
@@ -25,7 +26,7 @@ func (test *DepositDataSpecTest) Run(t *testing.T) {
 	validatorPK, _ := hex.DecodeString(test.ValidatorPK)
 	withdrawalCredentials, _ := hex.DecodeString(test.WithdrawalCredentials)
 
-	r, _, err := types.GenerateETHDepositData(
+	r, _, err := testingutils.GenerateETHDepositData(
 		validatorPK,
 		withdrawalCredentials,
 		test.ForkVersion,
