+++ types/testingutils/eth1_deposit.go
@@ -1,8 +1,9 @@
-package types
+package testingutils
 
 import (
 	"github.com/attestantio/go-eth2-client/spec/phase0"
 	"github.com/pkg/errors"
+	"github.com/ssvlabs/ssv-spec/types"
 )
 
 // GenerateETHDepositData returns un-signed deposit data and deposit data root for signature
@@ -16,20 +17,20 @@ func GenerateETHDepositData(
 	ret := &phase0.DepositMessage{
 		PublicKey:             pk,
 		WithdrawalCredentials: withdrawalCredentials,
-		Amount:                phase0.Gwei(MaxEffectiveBalanceInGwei),
+		Amount:                phase0.Gwei(types.MaxEffectiveBalanceInGwei),
 	}
 
-	domainR, err := ComputeETHDomain(domain, fork, GenesisValidatorsRoot)
+	domainR, err := types.ComputeETHDomain(domain, fork, types.GenesisValidatorsRoot)
 	if err != nil {
 		return nil, nil, errors.Wrap(err, "could not compute deposit domain")
 	}
-	signingRoot, err := ComputeETHSigningRoot(ret, domainR)
+	signingRoot, err := types.ComputeETHSigningRoot(ret, domainR)
 	if err != nil {
 		return nil, nil, errors.Wrap(err, "could not compute deposit signing root")
 	}
 	return signingRoot[:], &phase0.DepositData{
 		PublicKey:             pk,
 		WithdrawalCredentials: withdrawalCredentials,
-		Amount:                phase0.Gwei(MaxEffectiveBalanceInGwei),
+		Amount:                phase0.Gwei(types.MaxEffectiveBalanceInGwei),
 	}, nil
 }
